+++ src/routes/map/+page.svelte
@@ -5,6 +5,8 @@
 <MetaTags
     title={`World Map — Bapharia`}
     description={`Interactive map for BLUE PROTOCOL. Find enemies, locations, quests, treasure chests, gathering spots, and more!`}
+    image="/UI/Map/MapImage/Cty002/UI_MapCty002.png"
+    bigImage
 />
 
 <slot />